sdk/pyproject.toml
[project]
name = "tokenbudget"
version = "0.1.0"
description = "Open source AI cost tracking for OpenAI and Anthropic"
readme = "README.md"
license = {text = "Apache-2.0"}
authors = [{name = "TokenBudget Contributors"}]
keywords = ["openai", "anthropic", "ai", "cost", "tracking",
            "observability", "llm", "tokens"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.9"
dependencies = [
    "httpx>=0.27.0",
]

[project.optional-dependencies]
openai = ["openai>=1.0.0"]
anthropic = ["anthropic>=0.20.0"]
bedrock = ["boto3>=1.28"]
langchain = ["langchain-core>=0.2"]
llamaindex = ["llama-index-core>=0.10"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-httpx>=0.30",
    "ruff>=0.5",
]

[project.urls]
Homepage = "https://tokenbudget.com"
Repository = "https://github.com/AIMasterLabs/tokenbudget"
Documentation = "https://docs.tokenbudget.com"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py39"
line-length = 100
